Manually optimise tail call in sort comparator using a loop
It took me way too long to figure out the correct structure for making
this function tail call optimisable in the first place. Given that's the
case, I'd rather use a for-loop because I think it'd take just as long
for someone to understand it.

For completeness, the TCO-able version would look something like this:

    function comparator(sort, thenBys) {
      const f = (next, rest, a, b) => {
        const cmp = next(a, b);
        return cmp !== 0 || rest.length === 0
          ? cmp
          : f(rest[0], rest.slice(1), a, b);
      };
      return (a, b) => f(sort, thenBys, a, b);
    }